Safari-611.1.10/LayoutTests/accessibility/mac/search-predicate-element-count-expected.txt

This tests that search predicate based element count results are accurate.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tableCount is 2
PASS linkCount is 3
PASS imageCount is 4
PASS inputCount is 5
PASS controlCount is 12
PASS headingCount is 6
PASS onscreenCount is 6
PASS offscreenCount is 4
PASS successfullyParsed is true

TEST COMPLETE
